dates script refactor
You can print out work weeks and full weeks for a given day.

Refactor the functions to share what is common among the two. The type
hints in python are a little weird, but I wanted to try it out.

Add a range separator option so you can change the symbol that is the
range. Most of the time this is likely going to be an emdash.