Update VS Code Plugin handling
Move extension list from an array in the vscode.sh file into a separate
text file. This is so it is much easier to get the list of extensions.
You can use this command to get the extensions and update the list

code-insiders --list-extensions > vscode-plugins.txt

It looks like VS Code keeps these sorted so it should be easy to
compare. Which is what vscode-sync.sh does. It diffs what you have in
your instance to what is in the plugin list. By default it will install
any plugins missing. You can add --full to the argument and have it also
uninstall plugins that aren't in the list.

We will see how this goes, but it feels more useful and easier to
maintain than what was there previously.