Symlink items from bin/ to ~/bin
Since I'm not symlinking ~/bin directly to bin/ I need to symlink each
item from bin/ into ~/bin. Not sure if I'm a huge fan of this but it
does allow for adding items to ~/bin without having the dotfiles repo
say they're untracked.

Try this out for a while to see how I feel about it.